Move all layout components to a seperate folder
